Move more functions to civiimport ImportParser
